+++ lib/src/widgets/mute_sound_button.dart
@@ -28,7 +28,7 @@ class MuteSoundButton extends StatelessWidget {
       return PlayerButton(
         size: responsive.ip(_.fullscreen.value ? 5 : 7),
         circle: false,
-        backgrounColor: Colors.transparent,
+        backgroundColor: Colors.transparent,
         iconColor: Colors.white,
         iconPath: iconPath,
         customIcon: customIcon,