+++ lib/src/widgets/player_button.dart
@@ -5,7 +5,7 @@ class PlayerButton extends StatelessWidget {
   final double size;
   final String iconPath;
   final VoidCallback onPressed;
-  final Color backgrounColor, iconColor;
+  final Color backgroundColor, iconColor;
   final bool circle;
   final Widget? customIcon;
 
@@ -15,32 +15,33 @@ class PlayerButton extends StatelessWidget {
     required this.iconPath,
     required this.onPressed,
     this.circle = true,
-    this.backgrounColor = Colors.white54,
+    this.backgroundColor = Colors.white54,
     this.iconColor = Colors.black,
     this.customIcon,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return CupertinoButton(
-      padding: EdgeInsets.zero,
-      minSize: 20,
+    return TextButton(
+      style: TextButton.styleFrom(minimumSize: Size(20, 20)),
+      //padding: EdgeInsets.zero,
+      //minSize: 20,
+      onPressed: onPressed,
       child: customIcon ??
           Container(
-            width: this.size,
-            height: this.size,
-            padding: EdgeInsets.all(this.size * 0.25),
+            width: size,
+            height: size,
+            padding: EdgeInsets.all(size * 0.25),
+            decoration: BoxDecoration(
+              color: backgroundColor,
+              shape: circle ? BoxShape.circle : BoxShape.rectangle,
+            ),
             child: Image.asset(
-              this.iconPath,
-              color: this.iconColor,
+              iconPath,
+              color: iconColor,
               package: 'flutter_meedu_videoplayer',
             ),
-            decoration: BoxDecoration(
-              color: this.backgrounColor,
-              shape: this.circle ? BoxShape.circle : BoxShape.rectangle,
-            ),
           ),
-      onPressed: this.onPressed,
     );
   }
 }